G9-api/UploaderService/upload/requirements.txt
cloudinary==1.32.0
Flask==2.2.2
Flask_Cors==3.0.10
pytest==7.2.2
python-dotenv==1.0.0
waitress==2.1.2
Werkzeug==2.2.3
WTForms==3.0.1
